+++ src/demo_helpers/ui/shortcuts_help.py
@@ -192,7 +192,7 @@ class ShortcutsHelpWindow:
     WINDOW_TITLE = "Micro Tracker 3 - Shortcuts"
     F1_KEYCODES = F1_KEYCODES
 
-    def __init__(self, offset_xy: tuple[int, int] = (40, 40), version: str = "1.8.1"):
+    def __init__(self, offset_xy: tuple[int, int] = (40, 40), version: str = "1.8.2"):
         self._offset_xy = offset_xy
         self._is_visible = False
         self._panel_bgr = _render_shortcuts_panel(version)